Rename `cogs._checks.py`
