+++ cogs/archive.py
@@ -6,7 +6,7 @@ import re
 import discord
 from discord.ext import commands
 
-from cogs._checks import Checks
+from cogs._command_checks import Checks
 from cogs._utils import TeXBotApplicationContext, TeXBotAutocompleteContext, TeXBotCog
 from exceptions import BaseDoesNotExistError, UserNotInCSSDiscordServer
