+++ cogs/edit_message.py
@@ -5,7 +5,7 @@ import re
 import discord
 from discord.ext import commands
 
-from cogs._checks import Checks
+from cogs._command_checks import Checks
 from cogs._utils import TeXBotApplicationContext, TeXBotAutocompleteContext, TeXBotCog
 from exceptions import BaseDoesNotExistError, UserNotInCSSDiscordServer
